Start of main cleanup + random
Changed how files are read. Now it does not require a number of bodies at the beginning of the file. Console line arguments are instead used. If they don't exist, default values are used. The graphics and file reading can be turned off with booleans through code or console lines. Without file, planets are randomly generated.